+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## 0.1.1-dev
+
+- Add field for path to cwebp binary(optional)
+- Remove cwebp binaries from package
+
 ## 0.1.0-dev
 
 - Add feature: Convert .jpg and .png images to .webp